refactor components more
