+++ web/apps/playground/src/components/EditorPanel/index.ts
@@ -1,0 +1,1 @@
+export * from "./EditorPanel";
