--- web/apps/playground/src/components/PlaygroundPreview/index.ts
@@ -1,1 +1,0 @@
-export { PlaygroundPreview } from "./PlaygroundPreview";
